Replace BytesRefHash and clean up alternative implementations
