+++ server/src/main/java/org/opensearch/common/util/BytesRefHash.java
@@ -6,179 +6,291 @@
  * compatible open source license.
  */
 
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-/*
- * Modifications Copyright OpenSearch Contributors. See
- * GitHub history for details.
- */
-
 package org.opensearch.common.util;
 
+import net.openhft.hashing.LongHashFunction;
 import org.apache.lucene.util.BytesRef;
+import org.opensearch.common.Numbers;
 import org.opensearch.common.lease.Releasable;
 import org.opensearch.common.lease.Releasables;
 import org.opensearch.core.common.util.ByteArray;
 
+import java.security.AccessController;
+import java.security.PrivilegedAction;
+
 /**
- *  Specialized hash table implementation similar to Lucene's BytesRefHash that maps
- *  BytesRef values to ids. Collisions are resolved with open addressing and linear
- *  probing, growth is smooth thanks to {@link BigArrays}, hashes are cached for faster
- *  re-hashing and capacity is always a multiple of 2 for faster identification of buckets.
- *  This class is not thread-safe.
+ * Specialized hash table implementation that maps a {@link BytesRef} key to a long ordinal.
  *
- *  @opensearch.internal
+ * <p>
+ * It uses a compact byte-packing strategy to encode the ordinal and fingerprint information
+ * in the hash table value. It makes lookups faster by short-circuiting expensive equality checks
+ * for keys that collide onto the same hash table slot.
+ *
+ * <p>
+ * This class is not thread-safe.
+ *
+ * @opensearch.internal
  */
-public final class BytesRefHash extends AbstractHash {
+public class BytesRefHash implements Releasable {
+    private static final LongHashFunction XX3 = AccessController.doPrivileged(
+        (PrivilegedAction<LongHashFunction>) () -> LongHashFunction.xx3(System.nanoTime())
+    );
 
-    private LongArray startOffsets;
-    private ByteArray bytes;
-    private IntArray hashes; // we cache hashes for faster re-hashing
-    private final BytesRef spare;
+    private static final long MAX_CAPACITY = 1L << 32;
+    private static final long DEFAULT_INITIAL_CAPACITY = 32;
+    private static final float DEFAULT_LOAD_FACTOR = 0.6f;
+    private static final Hasher DEFAULT_HASHER = key -> XX3.hashBytes(key.bytes, key.offset, key.length);
 
-    // Constructor with configurable capacity and default maximum load factor.
-    public BytesRefHash(long capacity, BigArrays bigArrays) {
-        this(capacity, DEFAULT_MAX_LOAD_FACTOR, bigArrays);
+    private static final long MASK_ORDINAL = 0x00000000FFFFFFFFL;  // extract ordinal
+    private static final long MASK_FINGERPRINT = 0xFFFFFFFF00000000L;  // extract fingerprint
+
+    /**
+     * Maximum load factor after which the capacity is doubled.
+     */
+    private final float loadFactor;
+
+    /**
+     * Calculates the hash of a {@link BytesRef} key.
+     */
+    private final Hasher hasher;
+
+    /**
+     * Utility class to allocate recyclable arrays.
+     */
+    private final BigArrays bigArrays;
+
+    /**
+     * Reusable BytesRef to read keys.
+     */
+    private final BytesRef scratch = new BytesRef();
+
+    /**
+     * Current capacity of the hash table. This must be a power of two so that the hash table slot
+     * can be identified quickly using bitmasks, thus avoiding expensive modulo or integer division.
+     */
+    private long capacity;
+
+    /**
+     * Bitmask to identify the hash table slot from a key's hash.
+     */
+    private long mask;
+
+    /**
+     * Size threshold after which the hash table needs to be doubled in capacity.
+     */
+    private long grow;
+
+    /**
+     * Current size of the hash table.
+     */
+    private long size;
+
+    /**
+     * Underlying array to store the hash table values.
+     *
+     * <p>
+     * Each hash table value (64-bit) uses the following byte packing strategy:
+     * <pre>
+     * |================================|================================|
+     * | Fingerprint                    | Ordinal                        |
+     * |--------------------------------|--------------------------------|
+     * | 32 bits                        | 32 bits                        |
+     * |================================|================================|
+     * </pre>
+     *
+     * <p>
+     * This allows us to encode and manipulate additional information in the hash table
+     * itself without having to look elsewhere in the memory, which is much slower.
+     *
+     * <p>
+     * Terminology: <code>table[index] = value = (fingerprint | ordinal)</code>
+     */
+    private LongArray table;
+
+    /**
+     * Underlying array to store the starting offsets of keys.
+     *
+     * <p>
+     * Terminology:
+     * <pre>
+     *   offsets[ordinal] = starting offset (inclusive)
+     *   offsets[ordinal + 1] = ending offset (exclusive)
+     * </pre>
+     */
+    private LongArray offsets;
+
+    /**
+     * Underlying byte array to store the keys.
+     *
+     * <p>
+     * Terminology: <code>keys[start...end] = key</code>
+     */
+    private ByteArray keys;
+
+    public BytesRefHash(final BigArrays bigArrays) {
+        this(DEFAULT_INITIAL_CAPACITY, DEFAULT_LOAD_FACTOR, DEFAULT_HASHER, bigArrays);
     }
 
-    // Constructor with configurable capacity and load factor.
-    public BytesRefHash(long capacity, float maxLoadFactor, BigArrays bigArrays) {
-        super(capacity, maxLoadFactor, bigArrays);
-        startOffsets = bigArrays.newLongArray(capacity + 1, false);
-        startOffsets.set(0, 0);
-        bytes = bigArrays.newByteArray(capacity * 3, false);
-        hashes = bigArrays.newIntArray(capacity, false);
-        spare = new BytesRef();
+    public BytesRefHash(final long initialCapacity, final BigArrays bigArrays) {
+        this(initialCapacity, DEFAULT_LOAD_FACTOR, DEFAULT_HASHER, bigArrays);
     }
 
-    // BytesRef has a weak hashCode function so we try to improve it by rehashing using Murmur3
-    // Feel free to remove rehashing if BytesRef gets a better hash function
-    private static int rehash(int hash) {
-        return BitMixer.mix32(hash);
+    public BytesRefHash(final long initialCapacity, final float loadFactor, final BigArrays bigArrays) {
+        this(initialCapacity, loadFactor, DEFAULT_HASHER, bigArrays);
+    }
+
+    public BytesRefHash(final long initialCapacity, final float loadFactor, final Hasher hasher, final BigArrays bigArrays) {
+        assert initialCapacity > 0 : "initial capacity must be greater than 0";
+        assert loadFactor > 0 && loadFactor < 1 : "load factor must be between 0 and 1";
+
+        this.loadFactor = loadFactor;
+        this.hasher = hasher;
+        this.bigArrays = bigArrays;
+
+        capacity = Numbers.nextPowerOfTwo((long) (initialCapacity / loadFactor));
+        assert capacity <= MAX_CAPACITY : "required capacity too large";
+        mask = capacity - 1;
+        size = 0;
+        grow = (long) (capacity * loadFactor);
+
+        table = bigArrays.newLongArray(capacity, false);
+        table.fill(0, capacity, -1);
+        offsets = bigArrays.newLongArray(initialCapacity + 1, false);
+        offsets.set(0, 0);
+        keys = bigArrays.newByteArray(initialCapacity * 3, false);
     }
 
     /**
-     * Return the key at <code>0 &lt;= index &lt;= capacity()</code>. The result is undefined if the slot is unused.
-     * <p>Beware that the content of the {@link BytesRef} may become invalid as soon as {@link #close()} is called</p>
+     * Adds the given key to the hash table and returns its ordinal.
+     * If the key exists already, it returns (-1 - ordinal).
      */
-    public BytesRef get(long id, BytesRef dest) {
-        final long startOffset = startOffsets.get(id);
-        final int length = (int) (startOffsets.get(id + 1) - startOffset);
-        bytes.get(startOffset, length, dest);
+    public long add(final BytesRef key) {
+        final long hash = hasher.hash(key);
+        final long fingerprint = hash & MASK_FINGERPRINT;
+
+        for (long idx = hash & mask, value, ordinal;; idx = (idx + 1) & mask) {
+            if ((value = table.get(idx)) == -1) {
+                final long val = fingerprint | size;
+                if (size >= grow) {
+                    growAndInsert(hash, val);
+                } else {
+                    table.set(idx, val);
+                }
+                return append(key);
+            } else if (((value & MASK_FINGERPRINT) == fingerprint) && key.bytesEquals(get(ordinal = (value & MASK_ORDINAL), scratch))) {
+                return -1 - ordinal;
+            }
+        }
+    }
+
+    /**
+     * Returns the ordinal associated with the given key, or -1 if the key doesn't exist.
+     *
+     * <p>
+     * Using the 64-bit hash value, up to 32 least significant bits (LSB) are used to identify the
+     * home slot in the hash table, and an additional 32 bits are used to identify the fingerprint.
+     * The fingerprint further increases the entropy and reduces the number of false lookups in the
+     * keys' table during equality checks, which is expensive.
+     *
+     * <p>
+     * Total entropy bits = 32 + log2(capacity)
+     *
+     * <p>
+     * Linear probing starts from the home slot, until a match or an empty slot is found.
+     * Values are first checked using their fingerprint (to reduce false positives), then verified
+     * in the keys' table using an equality check.
+     */
+    public long find(final BytesRef key) {
+        final long hash = hasher.hash(key);
+        final long fingerprint = hash & MASK_FINGERPRINT;
+
+        for (long idx = hash & mask, value, ordinal;; idx = (idx + 1) & mask) {
+            if ((value = table.get(idx)) == -1) {
+                return -1;
+            } else if (((value & MASK_FINGERPRINT) == fingerprint) && key.bytesEquals(get(ordinal = (value & MASK_ORDINAL), scratch))) {
+                return ordinal;
+            }
+        }
+    }
+
+    /**
+     * Returns the key associated with the given ordinal.
+     * The result is undefined for an unused ordinal.
+     *
+     * <p>
+     * Beware that the content of the {@link BytesRef} may become invalid as soon as {@link #close()} is called
+     */
+    public BytesRef get(final long ordinal, final BytesRef dest) {
+        final long start = offsets.get(ordinal);
+        final int length = (int) (offsets.get(ordinal + 1) - start);
+        keys.get(start, length, dest);
         return dest;
     }
 
     /**
-     * Get the id associated with <code>key</code>
+     * Returns the number of mappings in this hash table.
      */
-    public long find(BytesRef key, int code) {
-        final long slot = slot(rehash(code), mask);
-        for (long index = slot;; index = nextSlot(index, mask)) {
-            final long id = id(index);
-            if (id == -1L || key.bytesEquals(get(id, spare))) {
-                return id;
-            }
-        }
+    public long size() {
+        return size;
     }
 
-    /** Sugar for {@link #find(BytesRef, int) find(key, key.hashCode()} */
-    public long find(BytesRef key) {
-        return find(key, key.hashCode());
+    /**
+     * Appends the key in the keys' and offsets' tables.
+     */
+    private long append(final BytesRef key) {
+        final long start = offsets.get(size);
+        final long end = start + key.length;
+        offsets = bigArrays.grow(offsets, size + 2);
+        offsets.set(size + 1, end);
+        keys = bigArrays.grow(keys, end);
+        keys.set(start, key.bytes, key.offset, key.length);
+        return size++;
     }
 
-    private long set(BytesRef key, int code, long id) {
-        assert rehash(key.hashCode()) == code;
-        assert size < maxSize;
-        final long slot = slot(code, mask);
-        for (long index = slot;; index = nextSlot(index, mask)) {
-            final long curId = id(index);
-            if (curId == -1) { // means unset
-                id(index, id);
-                append(id, key, code);
-                ++size;
-                return id;
-            } else if (key.bytesEquals(get(curId, spare))) {
-                return -1 - curId;
-            }
-        }
-    }
+    /**
+     * Grows the hash table by doubling its capacity, inserting the provided value,
+     * and reinserting the previous values at their updated slots.
+     */
+    private void growAndInsert(final long hash, final long value) {
+        // Ensure that the hash table doesn't grow too large.
+        // This implicitly also ensures that the ordinals are no larger than 2^32, thus,
+        // preventing them from polluting the fingerprint bits in the hash table values.
+        assert capacity < MAX_CAPACITY : "hash table already at the max capacity";
 
-    private void append(long id, BytesRef key, int code) {
-        assert size == id;
-        final long startOffset = startOffsets.get(size);
-        bytes = bigArrays.grow(bytes, startOffset + key.length);
-        bytes.set(startOffset, key.bytes, key.offset, key.length);
-        startOffsets = bigArrays.grow(startOffsets, size + 2);
-        startOffsets.set(size + 1, startOffset + key.length);
-        hashes = bigArrays.grow(hashes, id + 1);
-        hashes.set(id, code);
-    }
+        capacity <<= 1;
+        mask = capacity - 1;
+        grow = (long) (capacity * loadFactor);
+        table = bigArrays.grow(table, capacity);
+        table.fill(0, capacity, -1);
+        table.set(hash & mask, value);
 
-    private boolean assertConsistent(long id, int code) {
-        get(id, spare);
-        return rehash(spare.hashCode()) == code;
-    }
-
-    private void reset(int code, long id) {
-        assert assertConsistent(id, code);
-        final long slot = slot(code, mask);
-        for (long index = slot;; index = nextSlot(index, mask)) {
-            final long curId = id(index);
-            if (curId == -1) { // means unset
-                id(index, id);
-                break;
-            }
+        for (long ordinal = 0; ordinal < size; ordinal++) {
+            reinsert(ordinal, hasher.hash(get(ordinal, scratch)));
         }
     }
 
     /**
-     * Try to add <code>key</code>. Return its newly allocated id if it wasn't in the hash table yet, or <code>-1-id</code>
-     * if it was already present in the hash table.
+     * Reinserts the hash table value for an existing key stored at the given ordinal.
      */
-    public long add(BytesRef key, int code) {
-        if (size >= maxSize) {
-            assert size == maxSize;
-            grow();
+    private void reinsert(final long ordinal, final long hash) {
+        for (long idx = hash & mask;; idx = (idx + 1) & mask) {
+            if (table.get(idx) == -1) {
+                table.set(idx, (hash & MASK_FINGERPRINT) | ordinal);
+                return;
+            }
         }
-        assert size < maxSize;
-        return set(key, rehash(code), size);
-    }
-
-    /** Sugar to {@link #add(BytesRef, int) add(key, key.hashCode()}. */
-    public long add(BytesRef key) {
-        return add(key, key.hashCode());
-    }
-
-    @Override
-    protected void removeAndAdd(long index) {
-        final long id = id(index, -1);
-        assert id >= 0;
-        final int code = hashes.get(id);
-        reset(code, id);
     }
 
     @Override
     public void close() {
-        try (Releasable releasable = Releasables.wrap(bytes, hashes, startOffsets)) {
-            super.close();
-        }
+        Releasables.close(table, offsets, keys);
     }
 
+    /**
+     * Hasher calculates the hash of a {@link BytesRef} key.
+     */
+    @FunctionalInterface
+    public interface Hasher {
+        long hash(BytesRef key);
+    }
 }
