+++ server/src/test/java/org/opensearch/common/util/BytesRefHashTests.java
@@ -33,6 +33,7 @@
 package org.opensearch.common.util;
 
 import org.apache.lucene.tests.util.TestUtil;
+import net.openhft.hashing.LongHashFunction;
 import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.BytesRefBuilder;
 import org.opensearch.common.settings.Settings;
@@ -44,6 +45,7 @@ import java.util.HashSet;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
+import java.util.stream.Stream;
 
 public class BytesRefHashTests extends OpenSearchTestCase {
 
@@ -57,9 +59,13 @@ public class BytesRefHashTests extends OpenSearchTestCase {
         if (hash != null) {
             hash.close();
         }
-        // Test high load factors to make sure that collision resolution works fine
-        final float maxLoadFactor = 0.6f + randomFloat() * 0.39f;
-        hash = new BytesRefHash(randomIntBetween(0, 100), maxLoadFactor, randomBigArrays());
+        LongHashFunction hasher = LongHashFunction.xx3(randomLong());
+        hash = new BytesRefHash(
+            randomIntBetween(1, 100),      // random capacity
+            0.6f + randomFloat() * 0.39f,  // random load factor to verify collision resolution
+            key -> hasher.hashBytes(key.bytes, key.offset, key.length),
+            randomBigArrays()
+        );
     }
 
     @Override
@@ -68,39 +74,34 @@ public class BytesRefHashTests extends OpenSearchTestCase {
         newHash();
     }
 
-    public void testDuel() {
-        final int len = randomIntBetween(1, 100000);
-        final BytesRef[] values = new BytesRef[len];
-        for (int i = 0; i < values.length; ++i) {
-            values[i] = new BytesRef(randomAlphaOfLength(5));
-        }
-        final Map<BytesRef, Integer> valueToId = new HashMap<>();
-        final BytesRef[] idToValue = new BytesRef[values.length];
-        final int iters = randomInt(1000000);
-        for (int i = 0; i < iters; ++i) {
-            final BytesRef value = randomFrom(values);
-            if (valueToId.containsKey(value)) {
-                assertEquals(-1 - valueToId.get(value), hash.add(value, value.hashCode()));
+    public void testFuzzy() {
+        Map<BytesRef, Long> reference = new HashMap<>();
+        BytesRef[] keys = Stream.generate(() -> new BytesRef(randomAlphaOfLength(20)))
+            .limit(randomIntBetween(1000, 2000))
+            .toArray(BytesRef[]::new);
+
+        // Verify the behaviour of "add" and "find".
+        for (int i = 0; i < keys.length * 10; i++) {
+            BytesRef key = keys[i % keys.length];
+            if (reference.containsKey(key)) {
+                long expectedOrdinal = reference.get(key);
+                assertEquals(-1 - expectedOrdinal, hash.add(key));
+                assertEquals(expectedOrdinal, hash.find(key));
             } else {
-                assertEquals(valueToId.size(), hash.add(value, value.hashCode()));
-                idToValue[valueToId.size()] = value;
-                valueToId.put(value, valueToId.size());
+                assertEquals(-1, hash.find(key));
+                reference.put(key, (long) reference.size());
+                assertEquals((long) reference.get(key), hash.add(key));
             }
         }
 
-        assertEquals(valueToId.size(), hash.size());
-        for (final var next : valueToId.entrySet()) {
-            assertEquals(next.getValue().longValue(), hash.find(next.getKey(), next.getKey().hashCode()));
+        // Verify the behaviour of "get".
+        BytesRef scratch = new BytesRef();
+        for (Map.Entry<BytesRef, Long> entry : reference.entrySet()) {
+            assertEquals(entry.getKey(), hash.get(entry.getValue(), scratch));
         }
 
-        for (long i = 0; i < hash.capacity(); ++i) {
-            final long id = hash.id(i);
-            BytesRef spare = new BytesRef();
-            if (id >= 0) {
-                hash.get(id, spare);
-                assertEquals(idToValue[(int) id], spare);
-            }
-        }
+        // Verify the behaviour of "size".
+        assertEquals(reference.size(), hash.size());
         hash.close();
     }
